refactor: 커스텀 훅으로 정렬/필터로직 분리
